Add files via upload
This folder contains 4 classes that I made so that my servlets may utilize them for the registration process, logging in, creating a survey as a logged in and verified user and storing the survey into the database. I also added Meri's servlet to display the survey that was just created by a registered and verified user.